Handle hot-loaded GrandCentral::Store
We were previously generating code that, when editing the store, would
create a whole new store when store.rb was hot-loaded, which would break
not only GC::DevTools, but also its integration with the app itself.
This commit allows the generated code for the store to be edited without
breaking these things.